CST435_MULTICONTAINER/requirements.txt
grpcio
grpcio-tools
cryptography
langdetect
vaderSentiment
deep-translator